+++ R/melt.r
@@ -3,13 +3,16 @@
 #' @param ...       Array[s] or data.frame[s] to be melted
 #' @param dimnames  List of names along the dimensions
 #' @param na_rm     Remove rows with NAs
+#' @return   data.frame with 'value' (or object names if multiple) indexed by axes
 #' @export
 melt = function(..., dimnames=NULL, na_rm=TRUE) {
-    dots = list(...)
+    dots = lapply(named_dots(...), eval, envir=parent.frame())
     for (i in seq_along(dots)) {
         if (!is.null(dimnames))
             names(dimnames(dots[[i]])) = dimnames[1:length(dim(dots[[i]]))]
         dots[[i]] = melt_one(dots[[i]], na_rm=na_rm)
+        if (length(dots) > 1)
+            colnames(dots[[i]])[ncol(dots[[i]])] = names(dots)[i]
     }
 
     Reduce(function(a,b) merge(a,b,all=!na_rm), dots)