+++ tests/testthat/test-mask.r
@@ -16,3 +16,12 @@ test_that("mask", {
     expect_equal(Zref, mask(F, along=2))
     expect_equal(Zref, t(mask(F, along=1)))
 })
+
+test_that("NA is dropped without error", {
+    G = c("a", "b", NA, "a")
+    Gm = mask(G)
+    Gref = structure(c(TRUE, FALSE, FALSE, TRUE, FALSE, TRUE, FALSE, FALSE),
+                     .Dim = c(4L, 2L), .Dimnames = list(NULL, c("a", "b")))
+
+    expect_equal(Gm, Gref)
+})